JD's first python script test on Discovery
This script was designed to be run on an interactive node (Andes) on discovery just to test out some basic loading and saving. It loads a .mat timeseries, makes a string describing the timeseries dimensions, prints that string, then saves that string. Importantly, the script loads the .mat file directly from the home directory (e.g., /dartfs-hpc/rc/home/z/f00589z) and saves into a subfolder on the home directory. Initially tried to load from the scratch directory, but couldn’t figure out how to access it. Going to assume this won’t be a problem when switching to one of the non-interactive nodes, but worth noting for now.