simple shell script used to test job submission on Discovery
This is a simple shell script that submits a quick job to the Discovery cluster. It calls 'discovery_python_test_nonInteractive.py', which loads a timeseries (.mat file) from the scratch directory on Discovery, gets its dimensions, and saves the dimensions as a string in JD's home directory on Discovery. The purpose of this script was just to test the functionality of submitting a job on Discovery via shell script.