generate nifti files from .mat time series for listening and reading tasks
This script generates nifti files from the parseEPI-outputted EPI time series .mat files
(ending in "_interp.mat") from the listening and reading control tasks. The purpose of
this step is to provide easy-to-use inputs for generating pymvpa datasets that can then
be used for hyperalignment (or other pymvpa friendly things like decoding).